Integrate drum rompler with JSON (#305)
* feat: Add unbuffered file sample reader

* fix: Remove constexpr from AttackBufferingSampleReader and BufferedReader

* feat: Add SampleRepository for managing sample file paths

* feat: Integrate SampleRepository for file-based sample loading

* fix: Change attack_buffer_ram_ to AudioBlock type

* fix: Use begin() for AudioBlock data pointer

* fix: Value-initialize TempoEvent to prevent uninitialized warning

* fix: Reorder SequencerController members to fix -Wreorder

* fix: auto-format filesystem on mount failure

* feat: Implement SysEx file transfer for sample loading

* fix: Add missing include for midi::SystemExclusive

* fix: Include full StandardFileOps definition

* fix: Correct file path handling for VFS and robust null-termination

* fix: Re-enable manifest load error message

* chore: Log sample paths during manifest loading

* chore: Log each line read from manifest file

* fix: Handle odd byte lengths in file transfer sender

* docs: Document sample manifest file format

* refactor: Rename sample module and enable filesystem init

* feat: add jsmn json parser

* refactor: Load samples from config.json via ConfigurationManager

* feat: Implement JSON configuration manager for samples

* feat: Add bootup debug info: list files and print config

* fix: Use standard POSIX directory functions in list_files

* fix: Use dirent.h for directory function declarations

* feat: Add embedded default config fallback

* build: Replace pico_add_binary_data with xxd for config embedding

* feat: Add default sample configurations to config.default.json

* feat: Add color field to sample configuration

* feat: Update chunk counter in-place

* fix: Validate source file exists before sending

* feat: Log source and destination during file transfer

* chore: Update system startup, config limits, and MIDI constant

* chore: Adjust sample sender delay and reorder includes

* feat: Implement ACK/NACK for reliable MIDI SysEx file transfer

* feat: Implement reliable MIDI communication with ACK/NACK

* feat: Add percentage progress bar to sample_sender

* fix: Correct SysEx protocol for stripped start byte and size

* fix: Log SysEx command receive and send

* fix: Make SysEx manufacturer ID check robust and null-terminate filenames

* chore: remove no longer relevant rompler application

* fix: Correctly pass SysEx payload to protocol handler

* perf: Make config buffer static

* fix: Correct SysEx file transfer state on open failure

* feat: Add standard file operations for rompler

* feat: Add SysEx command to format filesystem

* fix: Add leading 0 to SysEx manufacturer ID in replies

* Refactor: Use official SysEx ID; maintain legacy compatibility

* style: Apply code formatting

* refactor: Centralize custom SysEx command handling in protocol

* fix: Resolve build errors by using sysex protocol enum tags

* feat: Query and display device firmware version

* fix: Resolve SysEx firmware request timeout, refactor protocol, and update sender

* refactor: Improve SysEx protocol command dispatch and reply handling

* fix: Prevent double-framing of SysEx messages

* chore: Log outgoing SysEx messages in hex

* chore: Suppress control change messages in sample sender log

* fix: Add detailed logging for MIDI SysEx message parsing

* fix: Enable SysEx messages for MIDI input

* debug: printf error if file is not found

* docs: Add SysEx command and file transfer protocol documentation

* refactor: Unify max path length constant to config::MAX_PATH_LENGTH

* fix: Add missing include for config.h

* refactor: Unify maximum path length constant

* refactor: Decouple sysex protocol from action execution

* fix: Add explicit state checks for SysEx file transfers

* fix: Reset state on invalid command during file transfer

* refactor: Isolate stateless commands in handle_no_body

* refactor: Extract SysEx manufacturer ID check and decompose packet handler

* fix: Sanitize file transfer paths to prevent traversal

* fix: Cancel active SysEx file write on new BeginFileWrite

* fix: Increase MIDI message timeout

* fix: Clear timeouts in sample sender to exit cleanly

* perf: Reduce MIDI timeouts and improve console output

* perf: Increase file transfer chunk size to 120 bytes

+++ drum/sysex/chunk.h
@@ -13,7 +13,7 @@ struct Chunk {
   typedef etl::array<uint8_t, MIDI::SysExMaxSize> Data;
 
   // Copy bytes to make sure we have ownership.
-  constexpr Chunk(const uint8_t *bytes, const uint8_t count)
+  constexpr Chunk(const uint8_t *bytes, const size_t count)
       : data(copy_bytes(bytes, count)), count(count) {
   }
